pyproject.toml

[tool.poetry]
name = "lti1p3platform"
version = "0.0.7"
description = "LTI 1.3 Platform implementation"
authors = ["Jun Tu <jun@openlearning.com>"]

[tool.poetry.dependencies]
python = "^3.7.2"
PyJWT = "1.7.1"
requests = {version = "2.24.0", extras = ["security"]}
jwcrypto = "^1.4.2"
Jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]
mypy = "^1.1.1"
pytest = "^7.2.2"
pylint = "^2.17.1"
types-PyJWT = "^1.7.1"
types-requests = "^2.28.11"
Django = "^3.1"
fastapi = "^0.95.0"
Jinja2 = "^3.1.2"
django-stubs = "^1.16.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
